+++ crates/frui_widgets/src/boxes.rs
@@ -3,11 +3,12 @@ use frui::prelude::*;
 use crate::BoxLayoutData;
 
 pub trait ChildWidget {
-    type ChildType<'a>: Widget where Self: 'a;
+    type ChildType<'a>: Widget
+    where
+        Self: 'a;
 
     fn child<'w>(&'w self) -> &Self::ChildType<'w>;
 }
-
 
 #[derive(RenderWidget, Default, Builder)]
 pub struct ConstrainedBox<T: RenderWidget + Widget> {
@@ -39,6 +40,7 @@ impl<T: RenderWidget + Widget> RenderWidget for ConstrainedBox<T> {
     fn layout(&self, ctx: RenderContext<Self>, constraints: Constraints) -> Size {
         let constraints = self.constraints.enforce(constraints);
         let child_size = ctx.child(0).layout(constraints);
+
         if child_size != Size::ZERO {
             child_size
         } else {
@@ -78,7 +80,7 @@ impl<T: RenderWidget + Widget> RenderWidget for ConstrainedBox<T> {
             let width = self.child().get_max_intrinsic_width(height);
             assert!(width.is_infinite());
             if !self.get_constraints().has_infinite_width() {
-                self.get_constraints().constrain_width(Some(width))
+                self.get_constraints().constrain_width(width)
             } else {
                 width
             }
@@ -86,13 +88,14 @@ impl<T: RenderWidget + Widget> RenderWidget for ConstrainedBox<T> {
     }
 
     fn compute_max_intrinsic_height(&self, width: f64) -> f64 {
-        if self.get_constraints().has_bounded_height() && self.get_constraints().has_tight_height() {
+        if self.get_constraints().has_bounded_height() && self.get_constraints().has_tight_height()
+        {
             self.get_constraints().max_height
         } else {
             let height = self.child().get_max_intrinsic_height(width);
             assert!(height.is_infinite());
             if !self.get_constraints().has_infinite_height() {
-                self.get_constraints().constrain_height(Some(height))
+                self.get_constraints().constrain_height(height)
             } else {
                 height
             }
@@ -106,7 +109,7 @@ impl<T: RenderWidget + Widget> RenderWidget for ConstrainedBox<T> {
             let width = self.child().get_min_intrinsic_width(height);
             assert!(width.is_infinite());
             if !self.get_constraints().has_infinite_width() {
-                self.get_constraints().constrain_width(Some(width))
+                self.get_constraints().constrain_width(width)
             } else {
                 width
             }
@@ -114,13 +117,14 @@ impl<T: RenderWidget + Widget> RenderWidget for ConstrainedBox<T> {
     }
 
     fn compute_min_intrinsic_height(&self, width: f64) -> f64 {
-        if self.get_constraints().has_bounded_height() && self.get_constraints().has_tight_height() {
+        if self.get_constraints().has_bounded_height() && self.get_constraints().has_tight_height()
+        {
             self.get_constraints().min_height
         } else {
             let height = self.child().get_min_intrinsic_height(width);
             assert!(height.is_infinite());
             if !self.get_constraints().has_infinite_height() {
-                self.get_constraints().constrain_height(Some(height))
+                self.get_constraints().constrain_height(height)
             } else {
                 height
             }
@@ -130,7 +134,7 @@ impl<T: RenderWidget + Widget> RenderWidget for ConstrainedBox<T> {
 
 #[derive(RenderWidget, Builder)]
 pub struct UnconstrainedBox<T: RenderWidget + Widget> {
-    pub child: T
+    pub child: T,
 }
 
 impl<'a, T: RenderWidget + Widget> ParentData for UnconstrainedBox<T> {
@@ -173,14 +177,18 @@ impl SizedBox {
     pub fn from_size<T: RenderWidget + Widget>(child: T, size: Size) -> impl RenderWidget + Widget {
         ConstrainedBox {
             child,
-            constraints: Constraints::tight_for(Some(size.width), Some(size.height))
+            constraints: Constraints::new_tight_for(Some(size.width), Some(size.height)),
         }
     }
 
-    pub fn new<T: RenderWidget + Widget>(child: T, width: Option<f64>, height: Option<f64>) -> impl RenderWidget + Widget {
+    pub fn new<T: RenderWidget + Widget>(
+        child: T,
+        width: Option<f64>,
+        height: Option<f64>,
+    ) -> impl RenderWidget + Widget {
         ConstrainedBox {
             child,
-            constraints: Constraints::tight_for(width, height)
+            constraints: Constraints::new_tight_for(width, height),
         }
     }
 
@@ -191,7 +199,7 @@ impl SizedBox {
     pub fn shrink<T: RenderWidget + Widget>(child: T) -> impl RenderWidget + Widget {
         ConstrainedBox {
             child,
-            constraints: Constraints::tight_for(None, None)
+            constraints: Constraints::new_tight_for(None, None),
         }
     }
 }
